+++ @types/UserFormData.d.ts
@@ -1,3 +1,5 @@
+import AgentTeamType from "~/enums/AgentTeamType";
+
 export interface UserDataSchema {
   id: string,
   role: number
@@ -35,9 +37,19 @@ export interface UserDataSchema {
   contractPercentage: string
   contractIban: string
   contractBic: string
+  contractInitialPaymentMethod: string
   referenceAgent: string
+  referenceAgentData: string
   accountCreatedAt: string
   accountUpdatedAt: string
   accountActivatedAt: string
   accountVerifiedAt: string
+  contractSignedAt: string
+  signinLogs: any
+  contractFiles: any[]
+  account_status: string
+  files: any[]
+  gold: boolean
+  hasSubAgents: boolean
+  agentTeamType: typeof AgentTeamType
 }