+++ lang/it/it-tables.js
@@ -44,6 +44,7 @@ export default {
   'sign-logs-event': "Evento",
   'sign-logs-date': "Data",
   'commission-type': "Tipologia Provvigione",
+  'commissions-assigned': "Provvigioni attive",
   'curr-month-commissions': "Provvigioni attuali",
   'commission-percentage': "Percentuale applicata",
   'calc-date': "Data",
@@ -54,6 +55,8 @@ export default {
   'calc-interest-recapitalized': "Int. Ricapitalizzato",
   'calc-interest-collected': "Int. Riscosso",
   'calc-brite': "Brite",
+  'calc-brite-partial': "Brite Parziali",
   'calc-movement-type': "Tipo Movimento",
   'calc-amount': "Valore",
+  'clients-number': "Clienti"
 }