+++ plugins/enums.ts
@@ -1,4 +1,6 @@
 import AccountStatuses from '../enums/AccountStatuses'
+import AgentTeamType from '../enums/AgentTeamType'
+import CommissionType from '../enums/CommissionType'
 import PersonTypes from '../enums/PersonTypes'
 import UserRoles from '../enums/UserRoles'
 import Genders from '../enums/Genders'
@@ -14,6 +16,8 @@ import {Plugin} from "@nuxt/types";
 
 export const enums = {
   AccountStatuses,
+  AgentTeamType,
+  CommissionType,
   CurrencyType,
   DocumentTypes,
   Genders,