+++ static/changelog.js
@@ -1,45 +1,61 @@
-module.exports = `<h2 id="v163smalltbasmall">v 1.6.3 - <small>TBA</small></h2>
+module.exports = `<h2 id="v164smalltbasmall">v 1.6.4 - <small>TBA</small></h2>
 <h3 id="aggiunte">Aggiunte</h3>
+<ul>
+<li>Campo "IBAN", "BIC" e "Note" nel report excel delle richieste</li>
+<li>Aggiunto campo "Tipologia Provvigioni squadra" in caso di agente con subagenti</li>
+<li>Possibilità di cancellazione di tutte le notifiche o di una intera categoria</li>
+<li>Aggiunto modale per visualizzare i clienti di un subagente</li>
+<li>Aggiunto colonna che mostra provvigioni attive per gli agenti</li>
+</ul>
+<h3 id="miglioramenti">Miglioramenti</h3>
+<ul>
+<li>Spostato impostazioni agente (anagrafica utente) in tab dedicata</li>
+<li>Aggiunta colonna Brite Parziali nella calcolatrice</li>
+<li>Ridotto durata animazione al cambio pagina</li>
+<li>Modale che mostra i clienti permette di vedere anche i sottoclienti di un subagente</li>
+</ul>
+<h2 id="v163small06022020small">v 1.6.3 - <small>06/02/2020</small></h2>
+<h3 id="aggiunte-1">Aggiunte</h3>
 <ul>
 <li>Pulsanti per scaricare il preventivo corrente</li>
 <li>Menù per il download delle ricevute nell'elenco movimenti</li>
 </ul>
 <h2 id="v162small28012020small">v 1.6.2 - <small>28/01/2020</small></h2>
-<h3 id="aggiunte-1">Aggiunte</h3>
+<h3 id="aggiunte-2">Aggiunte</h3>
 <ul>
 <li>Barra di ricerca nell'elenco degli utenti</li>
 <li>Barra di navigazione in mobile</li>
 <li>Calcolatrice per preventivi</li>
 </ul>
-<h3 id="miglioramenti">Miglioramenti</h3>
+<h3 id="miglioramenti-1">Miglioramenti</h3>
 <ul>
 <li>Lettura tabelle in mobile </li>
 </ul>
 <h2 id="v161small05012021small">v 1.6.1 - <small>05/01/2021</small></h2>
-<h3 id="aggiunte-2">Aggiunte</h3>
+<h3 id="aggiunte-3">Aggiunte</h3>
 <ul>
 <li>Transizioni al cambio delle pagine</li>
 <li>Aggiunta lista dei cambiamenti per ogni versione</li>
 <li>Email nel caso di notifica in app non letta entro 30 minuti</li>
 </ul>
-<h3 id="miglioramenti-1">Miglioramenti</h3>
+<h3 id="miglioramenti-2">Miglioramenti</h3>
 <ul>
 <li>Grafica splash screen iniziale</li>
 <li>Cambiato popup comunicazioni chat, da schermo intero a popup normale</li>
 </ul>
 <h2 id="v160small05012021small">v 1.6.0 - <small>05/01/2021</small></h2>
-<h3 id="aggiunte-3">Aggiunte</h3>
+<h3 id="aggiunte-4">Aggiunte</h3>
 <ul>
 <li>Sezione "Gold / Global club" in anagrafica utente</li>
 <li>Pulsante "Riscossione gold" in pagina richieste</li>
 <li>Modale per "Riscossione Gold" per riscossione interessi o prelievo oro fisico</li>
 </ul>
-<h3 id="miglioramenti-2">Miglioramenti</h3>
+<h3 id="miglioramenti-3">Miglioramenti</h3>
 <ul>
 <li>Grafico dashboard agenti e clienti - ora mostra anche 6 mesi futuri </li>
 </ul>
 <h2 id="v150small04012021small">v 1.5.0 - <small>04/01/2021</small></h2>
-<h3 id="aggiunte-4">Aggiunte</h3>
+<h3 id="aggiunte-5">Aggiunte</h3>
 <ul>
 <li>Campo "metodo pagamento iniziale"</li>
 <li>Campo "investimento iniziale in oro"</li>
@@ -48,7 +64,7 @@ module.exports = `<h2 id="v163smalltbasmall">v 1.6.3 - <small>TBA</small></h2>
 <li>Pulsante per scaricare la contabile di un versamento</li>
 <li>Pulsante per scaricare il report dei bonifici da effettuare</li>
 </ul>
-<h3 id="miglioramenti-3">Miglioramenti</h3>
+<h3 id="miglioramenti-4">Miglioramenti</h3>
 <ul>
 <li>Corretta validazione tipologia utente in anagrafica</li>
 <li>Aggiornata la grafica del menu laterale e barra di navigazione</li>